Support for datetime in response dictionaries
When we make_json_response, it is now allowed to have datetime.datetime
objects, which are serialized in ISO 8601 datetime format.

Note that such use must be consistent across resources in collection
(e.g. you should use same timestamp representation for all resources
in the returned collection), or else sorting might get confused.

Also, while we are there, we add a newline at the end of pretty-printed
JSON.

Change-Id: Idca50114042460b89a7865903f6d2314484133ec